+++ Inferno/inferno.package.json
@@ -3,5 +3,6 @@
     "Inferno/third-party/hydrate",
     "Inferno/third-party/index"
   ],
-  "output": "inferno-library.package.js"
+  "output": "inferno-library.package.js",
+  "types": "Inferno/Types.d.ts"
 }
