+++ package.json
@@ -7,10 +7,8 @@
       "name": "Dominic Gannaway",
       "email": "dg@domgan.com"
    },
-   "types": "types.d.ts",
    "files": [
-      "Inferno/*",
-      "types.d.ts"
+      "Inferno/*"
    ],
    "bugs": {
       "url": "https://github.com/infernojs/inferno/issues"
@@ -130,5 +128,6 @@
    },
    "dependencies": {
       "d3-scale-chromatic": "^1.3.3"
-   }
+   },
+  "types": "Inferno/Types.d.ts"
 }
