+++ tsconfig.json
@@ -72,6 +72,7 @@
     "outDir": "build",
     "sourceMap": false,
     "declaration": true,
+    "declarationDir": "./Inferno/types",
     "lib": [
       "es6",
       "dom"
